+++ subpackages/lib/source/odood/lib/server/exception.d
@@ -10,7 +10,13 @@ class ServerException : OdoodException
 }
 
 
-class ServerAlreadyRuningException : OdoodException
+class ServerAlreadyRuningException : ServerException
+{
+    mixin basicExceptionCtors;
+}
+
+
+class ServerCommandFailedException : ServerException
 {
     mixin basicExceptionCtors;
 }